Fixed overflow of numcalls and numsubroutines (exhibited by miranda 16k dataset).  Also fixed statistics displays (crashes, text overwriting text).
git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@2762 a8e31b63-e6ca-409b-a01b-2eec1b506579

+++ tools/src/dms/src/edu/uoregon/tau/dms/dss/ParaProfDataSession.java
@@ -366,8 +366,8 @@ public abstract class ParaProfDataSession  extends DataSession{
 	double exclSum = 0;
 	double inclSum = 0;
 	double maxInclSum = 0;
-	int callSum = 0;
-	int subrSum = 0;
+	double callSum = 0;
+	double subrSum = 0;
 
 
 	GlobalMapping allFunctions = this.getGlobalMapping();
